organize matrix headers
